+++ calculation/src/test/java/ee/qrental/calculation/application/service/strategy/PartialWeeksRateCalculationStrategyTest.java
@@ -6,6 +6,7 @@ import ee.qrental.transaction.domain.Constant;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.Month;
 
@@ -71,6 +72,6 @@ class PartialWeeksRateCalculationStrategyTest {
         final var result = instanceUnderTest.calculate(week);
 
         //then
-        assertEquals(1000L, result);
+        assertEquals(BigDecimal.valueOf(1000d), result);
     }
 }
